==> python_27_code/2_Simulation_Classification_Dataset_Train_Test_Split_py27.ipynb <==
# Description

The simulated raw data generated will be split into two datasets: train and test. Students will be provided the training set while the test set will be withheld to test their models on.

# Libraries

import pandas as pd
import sklearn
from sklearn.model_selection import train_test_split
from __future__ import print_function

# Versions

!python --version

items = [("Sklearn", sklearn), ("Pandas", pd)]
for item in items:
    print(item[0] + " version: " + str(item[1].__version__))

# Get Data

X = pd.read_hdf('/Users/davidziganto/Repositories/Synthetic_Dataset_Generation/data/py27/simulated_raw_data_py27.h5', 'table')

# Preprocess Data

y = X.pop('hired')

# Split Data

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=42)
training = pd.concat([X_train, y_train], axis=1)
training.reset_index(drop=True, inplace=True)

# Pickle X_test & y_test

import pickle

path = '/Users/davidziganto/Repositories/Synthetic_Dataset_Generation/pickle_files/py27/'

# Save X_test
with open(path + 'X_test_py27.pkl', 'wb') as picklefile:
    pickle.dump(X_test, picklefile)
    
# Save y_test
with open(path + 'y_test_py27.pkl', 'wb') as picklefile:
    pickle.dump(y_test, picklefile)

# Write Training Data To Disk

training.to_hdf('/Users/davidziganto/Repositories/Synthetic_Dataset_Generation/data/py27/simulated_raw_training_data_py27.h5',
                 'table',
                 mode='w',
                 append=True, 
                 complevel=9, 
                 complib='blosc',
                 fletcher32=True)
