==> python_27_code/3_Simulated_Classification_Dataset_Corruption_py27.ipynb <==
# Description

---

### Introduction
There are numerous ways data can be messy or corrupted. Sometimes data is missing. Sometimes data entry errors occur. And sometimes anomalous data presents itself. These are just a few of the ways data can be messy. Proper munging is required to procure optimal results. Therefore, to challenge students, I am simulating several of these data issues that are common in the wild. The specifics can be found in the next section.

### Corruption Process
There are three key corruption processes that will be utilized. First, outliers related to age will be introduced. It should be unclear to someone who doesn't know the generative process whether these are data entry errors or true outliers. Secondly, obvious data entry errors are introduced. For example, we know GPA should be in the range of 1.0 to 4.0 but there is a case where GPA = 4.21. Lastly, missing values and an incorrect data type are introduced. I removed certain values that were close to the mean for a given feature. The assumption is that novice data scientists will impute with the mean, hence the choice. See below for details.

---

# Libraries

%matplotlib inline
import matplotlib
import matplotlib.pyplot as plt
import numpy as np
import pandas as pd
import seaborn as sns
from __future__ import print_function

# Versions

!python --version

items = [("Numpy", np), ("Pandas", pd), ("Matplotlib", matplotlib), ("Seaborn", sns)]
for item in items:
    print(item[0] + " version: " + str(item[1].__version__))

# Get Data

# Read from disk
data = pd.read_hdf('/Users/davidziganto/Repositories/Synthetic_Dataset_Generation/data/py27/simulated_raw_training_data_py27.h5', 'table')

# DF Details

data.head()

data.describe()

# Corruption Process

Three key areas will be addressed:
1. Adding outliers
2. Simulating data entry errors
3. Generating missing & inf values

First, please note that I considered replacing many of the numeric features with strings. For example, changing 0 and 1 in *gender* to female and male. The code to do this is trivial and is found below. The problem is that assessing students who choose different mapping schemes from string to number makes assessment nearly impossible. Therefore, I decided to forgo this method.

# Considered changing numbers to text but would make scoring difficult bc how numbers are coded matters, 
# so decided to avoid altogether. Will include dictionary instead.

# Change degree to strings
#data.degree.replace(range(4), ['NB','B','M','P'], inplace=True)
# Change gender to F/M
#data.gender.replace(range(2), ['F','M'], inplace=True)
# Change major to strings
#data.major.replace(range(8), ['AN','BI','BS','CH','EN','JO','MA','PS'], inplace=True)
# Change bootcamp to No/Yes
#data.bootcamp.replace(range(2), ['No','Yes'], inplace=True)
# Change blog to No/Yes
#data.blog.replace(range(2), ['No','Yes'], inplace=True)
# Change hirec to No/Yes
#data.hired.replace(range(2), ['No','Yes'], inplace=True)

### 1. Add Outliers

# Note: these changes do NOT impact the models but are here for future improvements

np.random.seed(25)

age_outlier_idx = np.where(data['age']==39)[0][np.random.choice(range(len(data[data['age']==39])), size=3)]
for i, _ in data[data['age']==39].iterrows():
    if i in age_outlier_idx:
        data.loc[i, 'age'] = int(np.random.choice(a=range(120,161), size=1))

### 2. Simulate Data Entry Error

# Note: these changes definitely affect models bc values are not possible!

np.random.seed(42)
input_error_idx = np.random.randint(low=0, high=len(data), size=3)

# GPA range is [1.0, 4.0] so add one that's too high (e.g. 4.21)
data.loc[input_error_idx[0], 'gpa'] = 4.21

# blogger var is [0, 1] but add value of 2
data.loc[input_error_idx[1], 'blogger'] = 2
data.loc[input_error_idx[2], 'blogger'] = 2

data.describe()

### 3. Generate Missing Values & Inf

np.random.seed(199)

# simulate missing data
data.loc[np.where(data['experience'] == 18)[0][0], 'experience'] = np.nan
data.loc[np.where(data['github'] == 10)[0][0], 'github'] = np.nan

# Randomly select index to corrupt
nonsense_idx = np.random.randint(low=0, high=len(data), size=1)

# add 'inf' data type
data.loc[nonsense_idx, 'articles'] = np.inf

#### Proof of Corruption: Nulls

data.isnull().sum()

data.iloc[np.where(data.isnull())[0]]

#### Proof of Corruption: Inf

np.isinf(data.articles).sum()

data.iloc[np.where(data==np.inf)[0]]

#### Proof of Corruption: Big Picture

data.describe()

# Write To Disk

# Write to disk
data.to_hdf('/Users/davidziganto/Repositories/Synthetic_Dataset_Generation/data/py27/simulated_messy_training_data_py27.h5',
            'table',
            mode='w',
            append=True, 
            complevel=9, 
            complib='blosc',
            fletcher32=True)
